+++ adrestus-network/src/main/java/io/Adrestus/p2p/kademlia/server/KademliaNodeServer.java
@@ -17,7 +17,6 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import java.io.Serializable;
-import java.math.BigInteger;
 
 
 @Getter
@@ -43,7 +42,7 @@ public class KademliaNodeServer<K extends Serializable, V extends Serializable> 
         this(null, port, factory);
     }
 
-    public synchronized void run(DHTKademliaNodeAPI<BigInteger, NettyConnectionInfo, K, V> dhtKademliaNodeAPI) throws InterruptedException {
+    public synchronized void run(DHTKademliaNodeAPI<Long, NettyConnectionInfo, K, V> dhtKademliaNodeAPI) throws InterruptedException {
         assert !running;
 
         this.bossGroup = new NioEventLoopGroup();
