+++ adrestus-network/src/main/java/io/Adrestus/p2p/kademlia/server/NettyKademliaMessageHandler.java
@@ -16,7 +16,6 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import java.io.Serializable;
-import java.math.BigInteger;
 import java.util.List;
 
 import static io.netty.handler.codec.http.HttpResponseStatus.OK;
@@ -25,9 +24,9 @@ import static io.netty.handler.codec.http.HttpResponseStatus.OK;
 public class NettyKademliaMessageHandler<K extends Serializable, V extends Serializable> extends SimpleChannelInboundHandler<FullHttpRequest> {
     private static Logger log = LoggerFactory.getLogger(NettyKademliaMessageHandler.class);
     private final NettyKademliaServerFilterChain<K, V> filterChain;
-    private final DHTKademliaNodeAPI<BigInteger, NettyConnectionInfo, K, V> dhtKademliaNodeAPI;
+    private final DHTKademliaNodeAPI<Long, NettyConnectionInfo, K, V> dhtKademliaNodeAPI;
 
-    public NettyKademliaMessageHandler(DHTKademliaNodeAPI<BigInteger, NettyConnectionInfo, K, V> dhtKademliaNodeAPI, NettyKademliaServerFilterChain<K, V> filterChain) {
+    public NettyKademliaMessageHandler(DHTKademliaNodeAPI<Long, NettyConnectionInfo, K, V> dhtKademliaNodeAPI, NettyKademliaServerFilterChain<K, V> filterChain) {
         this.filterChain = filterChain;
         this.dhtKademliaNodeAPI = dhtKademliaNodeAPI;
     }
