+++ adrestus-network/src/main/java/io/Adrestus/p2p/kademlia/server/NettyServerInitializer.java
@@ -6,10 +6,9 @@ import io.netty.channel.ChannelInboundHandler;
 import io.netty.channel.ChannelPipeline;
 
 import java.io.Serializable;
-import java.math.BigInteger;
 
 public interface NettyServerInitializer<K extends Serializable, V extends Serializable> extends ChannelInboundHandler {
-    void registerKademliaNode(DHTKademliaNodeAPI<BigInteger, NettyConnectionInfo, K, V> dhtKademliaNodeAPI);
+    void registerKademliaNode(DHTKademliaNodeAPI<Long, NettyConnectionInfo, K, V> dhtKademliaNodeAPI);
 
     void pipelineInitializer(ChannelPipeline pipeline);
 }
